Move fillDataGapsAndRoundCaps to components/chart
Now used in statistics-chart too, so move to common chart location.
Re-export in energy-chart-options.ts to minimise changes throughout energy cards.
