+++ src/panels/lovelace/cards/energy/common/energy-chart-options.ts
@@ -16,7 +16,6 @@ import {
   subDays,
 } from "date-fns";
 import type {
-  BarSeriesOption,
   CallbackDataParams,
   LineSeriesOption,
   TopLevelFormatterParams,
@@ -37,6 +36,8 @@ import { filterXSS } from "../../../../../common/util/xss";
 import type { StatisticPeriod } from "../../../../../data/recorder";
 import { getPeriodicAxisLabelConfig } from "../../../../../components/chart/axis-label";
 import { getSuggestedPeriod } from "../../../../../data/energy";
+
+export { fillDataGapsAndRoundCaps } from "../../../../../components/chart/round-caps";
 
 /**
  * Energy chart data point tuple:
@@ -278,72 +279,6 @@ function formatTooltip(
     footer += `<br><b>${formatTotal(sumNegative)}</b>`;
   }
   return values.length > 0 ? `${title}${values.join("<br>")}${footer}` : "";
-}
-
-export function fillDataGapsAndRoundCaps(datasets: BarSeriesOption[]) {
-  const buckets = Array.from(
-    new Set(
-      datasets
-        .map((dataset) =>
-          dataset.data!.map((datapoint) => Number(datapoint![0]))
-        )
-        .flat()
-    )
-  ).sort((a, b) => a - b);
-
-  // make sure all datasets have the same buckets
-  // otherwise the chart will render incorrectly in some cases
-  buckets.forEach((bucket, index) => {
-    const capRounded = {};
-    const capRoundedNegative = {};
-    for (let i = datasets.length - 1; i >= 0; i--) {
-      const dataPoint = datasets[i].data![index];
-      const item: any =
-        dataPoint && typeof dataPoint === "object" && "value" in dataPoint
-          ? dataPoint
-          : { value: dataPoint };
-      const x = item.value?.[0];
-      const stack = datasets[i].stack ?? "";
-      if (x === undefined) {
-        continue;
-      }
-      if (Number(x) !== bucket) {
-        datasets[i].data?.splice(index, 0, {
-          value: [bucket, 0],
-          itemStyle: {
-            borderWidth: 0,
-          },
-        });
-      } else if (item.value?.[1] === 0) {
-        // remove the border for zero values or it will be rendered
-        datasets[i].data![index] = {
-          ...item,
-          itemStyle: {
-            ...item.itemStyle,
-            borderWidth: 0,
-          },
-        };
-      } else if (!capRounded[stack] && item.value?.[1] > 0) {
-        datasets[i].data![index] = {
-          ...item,
-          itemStyle: {
-            ...item.itemStyle,
-            borderRadius: [4, 4, 0, 0],
-          },
-        };
-        capRounded[stack] = true;
-      } else if (!capRoundedNegative[stack] && item.value?.[1] < 0) {
-        datasets[i].data![index] = {
-          ...item,
-          itemStyle: {
-            ...item.itemStyle,
-            borderRadius: [0, 0, 4, 4],
-          },
-        };
-        capRoundedNegative[stack] = true;
-      }
-    }
-  });
 }
 
 function getDatapointX(datapoint: NonNullable<LineSeriesOption["data"]>[0]) {
